refactor components + fix submit button not close + fix bug when edit MaintenanceStatus
